Move validation test suite into a shared place
The test suite can be used in different ways to check the validation
functionality.
